services/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.12
pyyaml>=6.0.1
httpx>=0.27
aiohttp>=3.9
paho-mqtt>=1.6
websockets>=12.0
jinja2>=3.1
itsdangerous>=2.1
werkzeug>=3.0
